Establishes application behaviors and workflow
This adds the desired behaviors for the application UI - that is,
adding the event listeners for the buttons and implementing their
desired outcomes. "Generate" now populates the `TextView` with the
text from the `Generate` module, "Copy to Clipboard" copies the
current contents of the `TextView` to the system clipboard, and
`Clear` empties the `TextView`. With the wiring in place, we can
move on to the actual business logic of generating missions,
environments, and worlds.