Adds world generation logic
This adds the business logic to generate planets from a single
provided `StdGen`, which we split into eight different generators
so the same `StdGen` doesn't produce the same relative outcomes
for all generated parameters. This may be unnecessary, but it does
prevent possible issues surrounding the same generator picking the
same result each time it rolls the same die value.